% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution.R
\name{dilution}
\alias{dilution}
\title{Calculating dilutions}
\usage{
dilution(C1 = NA, V1 = NA, C2 = NA, V2 = NA)
}
\arguments{
\item{C1}{initial concentration}

\item{V1}{initial volume}

\item{C2}{desired concentration}

\item{V2}{desired volume}
}
\value{
The missing value needed for the dilution
}
\description{
Calculating dilutions
}
\examples{
dilution(C1 = 5, V1 = 10, C2 = 1)
}
