% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doorzetten.R
\name{doorzetten}
\alias{doorzetten}
\title{Hulp bij doorwerken}
\usage{
doorzetten(x)
}
\arguments{
\item{x}{moet 1 van e volgende character strings zijn: "ja", "nee", "bijna", "error"}
}
\value{
wat tekst om je te steunen of laten instorten
}
\description{
Hulp bij doorwerken
}
\examples{
doorzetten("ja")
doorzetten("error")
}
