% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gram_change.R
\name{gram_change}
\alias{gram_change}
\title{gram_changer}
\usage{
gram_change(x, unit, desired_unit)
}
\arguments{
\item{x}{An integer.}

\item{unit}{The unit of measurement used for the integer x.}

\item{desired_unit}{The unit of measurement you would like the integer to be changed into.}
}
\value{
A changed integer with the desired unit after
}
\description{
gram_changer
}
\examples{
gram_change(1205.24, "mg", "g")
}
