% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potter.R
\name{potter}
\alias{potter}
\title{potter}
\usage{
potter(x)
}
\arguments{
\item{x}{A number/integer ranging from 1 till 7. Can also be a string, if you remember a part of the title.}
}
\value{
A character string with information about which book the number is in the series
}
\description{
potter
}
\examples{
potter(4)
potter("goblet")
}
