% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trying.R
\name{project}
\alias{project}
\title{projecticum hulp}
\usage{
project(x)
}
\arguments{
\item{x}{The character string to file needed}
}
\value{
a few lines of genes and the umap
}
\description{
projecticum hulp
}
\examples{
\dontrun{
project("../dsfb2_workflows_exercises/Seurat_oefenen/justus/pbmc3k_filtered_gene_bc_matrices/filtered_gene_bc_matrices/hg19/")
}
}
