% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\title{files()}
\usage{
files(x, dir_path = NA)
}
\arguments{
\item{x}{either "kort" or "lang"}

\item{dir_path}{a path from which you want the list of folders}
}
\value{
A character string with information about working directory and a list with the folder structure
}
\description{
files()
}
\examples{
files("kort", "/tests/")
files("lang")
}
