% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potter.R
\name{potter}
\alias{potter}
\title{potter}
\usage{
potter(x)
}
\arguments{
\item{x}{A number/integer or character string}
}
\value{
A character string with information about which book the number or string is in the series
}
\description{
potter
}
\examples{
potter(4)
}
